Screenshots.js - Implement a name filter
This helps devs and users searching screenshots from another user or from an specific map.

The code is a bit slow, since it reads and builds html elements for all the ignored screenshots.

Probably this could be faster, but I don't know in detail how the backend of this works.

Update screenshots.js